UEFI and Ubuntu 14.04 support for chef-bach  - First attempt at updating shell scripts for Ubuntu 14.04  - Make cobbler recipe pass rubocop  - Remove hardcoded paths from gem installation recipes.  - Update bcpc::ufw recipe for 14.04 and rubocop compliance  - Update bcpc::apache-mirror for 14.04  - Update dependencies in metadata.rb and Berksfile to match shell scripts  - Add .chef_found* to chefignore  - Use upstream packages and cobbler 2.6  - Use os_version 'trusty' for ubuntu 14.04 image import  - Update enroll_cobbler.sh for cobbler 2.6  - Use Xenial HWE kernel for EFI installs; reorganize cobbler templates.  - Add 14.04 preseed, remove 12.04 preseed  - Correct whitespace on bcpc default attributes  - Fix vbox_create.sh for VM builds with EFI (again)  - Don't write empty strings to environment if no proxy specified  - automated_install.sh: Don't overwrite host's cluster dir in VM builds  - Override data disks in the stub environment for VM builds  - Move serial console kopts from enroll_cobbler.sh into bcpc::cobbler  - Use the -56 percona metapackage instead of the versionless one  - Updated quorum status check not to use mysql cli  - Fix disk handling for EFI VM builds  - Move percona repo to bcpc::mysql_client recipe  - Delete a couple of bogus methods from bcpc::hadoop utils.rb  - Fix mysql logging  - Improve rubocop compliance on mysql.rb  - Rubocop-ize bcpc-hadoop::oozie  - Move oozie recipes from shell scripts to database cookbook providers  - Move mysql connection info into helpers  - Don't create ipxe hardlink until cobbler sync runs.  - Straighten out tftp paths on bcpc::cobbler  - Delete (duplicate) defaults from Test-Laptop demo environment  - Grant global privileges to root@'%' in bcpc::mysql  - Correct root@'%' priveleges; memoize passwords in connection info

+++ build_bins.sh
@@ -43,7 +43,7 @@ mkdir -p $APT_REPO_BINS
 apt-get -y update
 
 # Install tools needed for packaging
-apt-get -y install git rubygems make pkg-config pbuilder python-mock python-configobj python-support cdbs python-all-dev python-stdeb libmysqlclient-dev libldap2-dev ruby-dev gcc patch rake ruby1.9.3 ruby1.9.1-dev python-pip python-setuptools dpkg-dev apt-utils haveged libtool autoconf automake autotools-dev unzip rsync autogen
+apt-get -y install git ruby make pkg-config pbuilder python-mock python-configobj python-support cdbs python-all-dev python-stdeb libmysqlclient-dev libldap2-dev ruby-dev gcc patch rake ruby1.9.3 ruby1.9.1-dev python-pip python-setuptools dpkg-dev apt-utils haveged libtool autoconf automake autotools-dev unzip rsync autogen
 
 # Install json gem first to avoid a too-new version being pulled in by other gems.
 if [[ -z `gem list --local json | grep json | cut -f1 -d" "` ]]; then
@@ -94,8 +94,8 @@ if ! [[ -f jce_policy-8.zip ]]; then
 fi
 FILES="jce_policy-8.zip $FILES"
 
-# Pull all the gems required for the cluster 
-for i in patron wmi-lite simple-graphite ruby-augeas; do
+# Pull all the (unversioned) gems required for the cluster 
+for i in patron wmi-lite simple-graphite ruby-augeas chef-rewind; do
   if ! [[ -f gems/${i}.gem ]]; then
     gem fetch ${i}
     ln -s ${i}-*.gem ${i}.gem || true
@@ -103,12 +103,12 @@ for i in patron wmi-lite simple-graphite ruby-augeas; do
   FILES="${i}*.gem $FILES"
 done
 
-# Get the Rubygem for zookeeper
-if ! [[ -f gems/zookeeper.gem ]]; then
-  gem fetch zookeeper -v 1.4.7
-  ln -s zookeeper-*.gem zookeeper.gem || true
+# Get the Rubygem for mysql2
+if ! [[ -f gems/mysql2.gem ]]; then
+  gem fetch mysql2 -v 0.4.4
+  ln -s mysql2-*.gem mysql2.gem || true
 fi
-FILES="zookeeper*.gem $FILES"
+FILES="mysql2*.gem $FILES"
 
 # Get the Rubygem for kerberos
 if ! [[ -f gems/rake-compiler.gem ]]; then
@@ -117,12 +117,26 @@ if ! [[ -f gems/rake-compiler.gem ]]; then
 fi
 FILES="rake-compiler*.gem $FILES"
 
-# Get the Rubygem for kerberos
+# Get the Rubygem for sequel
+if ! [[ -f gems/sequel.gem ]]; then
+  gem fetch sequel -v 4.36.0
+  ln -s sequel-*.gem sequel.gem || true
+fi
+FILES="sequel*.gem $FILES"
+
+# Get the Rubygem for rkerberos
 if ! [[ -f gems/rkerberos.gem ]]; then
   gem fetch rkerberos
   ln -s rkerberos*.gem rkerberos.gem || true
 fi
 FILES="rkerberos*.gem $FILES"
+
+# Get the Rubygem for webhdfs
+if ! [[ -f gems/webhdfs.gem ]]; then
+  gem fetch webhdfs -v 0.5.5
+  ln -s webhdfs-*.gem webhdfs.gem || true
+fi
+FILES="webhdfs*.gem $FILES"
 
 # Get Rubygem for zabbixapi
 if ! [[ -f gems/zabbixapi.gem ]]; then
@@ -131,12 +145,12 @@ if ! [[ -f gems/zabbixapi.gem ]]; then
 fi
 FILES="zabbix*.gem $FILES"
 
-# Get the Rubygem for webhdfs
-if ! [[ -f gems/webhdfs.gem ]]; then
-  gem fetch webhdfs -v 0.5.5
-  ln -s webhdfs-*.gem webhdfs.gem || true
+# Get the Rubygem for zookeeper
+if ! [[ -f gems/zookeeper.gem ]]; then
+  gem fetch zookeeper -v 1.4.7
+  ln -s zookeeper-*.gem zookeeper.gem || true
 fi
-FILES="webhdfs*.gem $FILES"
+FILES="zookeeper*.gem $FILES"
 
 # Fetch the cirros image for testing
 if ! [[ -f cirros-0.3.0-x86_64-disk.img ]]; then
@@ -146,15 +160,12 @@ if ! [[ -f cirros-0.3.0-x86_64-disk.img ]]; then
 fi
 FILES="cirros-0.3.0-x86_64-disk.img $FILES"
 
-# Grab the Ubuntu 12.04 installer image
-UBUNTU_IMAGE="ubuntu-12.04-mini.iso"
-
+# Grab the Ubuntu 14.04 installer image with a 4.4 "HWE" kernel from Xenial
+UBUNTU_IMAGE="ubuntu-14.04-hwe44-mini.iso"
 if ! [[ -f $UBUNTU_IMAGE ]]; then
-  # Download this ISO to get the latest kernel/X LTS stack installer
-  #$CURL -o $UBUNTU_IMAGE http://archive.ubuntu.com/ubuntu/dists/precise-updates/main/installer-amd64/current/images/raring-netboot/mini.iso
-  while ! $(file $UBUNTU_IMAGE | grep -qE '(x86 boot sector)|(ISO 9660 CD-ROM)'); do
-    $CURL -o $UBUNTU_IMAGE http://archive.ubuntu.com/ubuntu/dists/precise-updates/main/installer-amd64/20101020ubuntu136.21/images/netboot/mini.iso
-  done
+    while ! $(file $UBUNTU_IMAGE | grep -qE '(x86 boot sector)|(ISO 9660 CD-ROM)'); do
+	$CURL -o $UBUNTU_IMAGE http://archive.ubuntu.com/ubuntu/dists/trusty-updates/main/installer-amd64/current/images/xenial-netboot/mini.iso
+    done
 fi
 FILES="$UBUNTU_IMAGE $FILES"
 
@@ -329,15 +340,24 @@ chmod -R 755 .
 
 ####################
 # generate Pypi repo
-
-# Wheel installs require setuptools >= 0.8 for dist-info support.
-# can then follow http://askubuntu.com/questions/399446
-# but can't upgrade setuptools first as:
-# "/usr/bin/pip install: error: no such option: --no-use-wheel"
 if ! hash dir2pi; then
-  /usr/bin/pip install pip2pi || /bin/true
-  /usr/local/bin/pip install setuptools --no-use-wheel --upgrade
-  /usr/local/bin/pip install pip2pi
+    PIP_VERSION=`pip --version | perl -nle 'm/pip\s+([\d\.]+)/; print $1'`
+    
+    # If we have an ancient pip, upgrade it before getting pip2pi.
+    if ruby -e "exit 1 if Gem::Version.new('$PIP_VERSION') > \ 
+                          Gem::Version.new('1.5.4')"
+    then
+	# Wheel installs require setuptools >= 0.8 for dist-info support.
+	# can then follow http://askubuntu.com/questions/399446
+	# but can't upgrade setuptools first as:
+	# "/usr/bin/pip install: error: no such option: --no-use-wheel"
+	echo "Upgrading pip before pip2pi install"
+	/usr/bin/pip install pip2pi || /bin/true
+	/usr/local/bin/pip install setuptools --no-use-wheel --upgrade
+	/usr/local/bin/pip install pip2pi
+    else
+	/usr/bin/pip install pip2pi
+    fi	 
 fi
 
 dir2pi python
@@ -356,3 +376,4 @@ fi
 gem generate_index --legacy
 
 popd
+
